More idiomatic code
Phrase is generally created from query_from_sentence, which tokenizes the sentence and creates all the words of choice. However, the Phrase constructor would then tokenize the sentence again, which is a waste of time. This has been changed. Additionally, we return just the labels and words from the doc in Parser.tokenize_sentence, but the doc might have useful information. We now return the doc as part of a Sentence class.