Simplify CFG parsing
This commit removes most string indexing in the CFG parser. Instead, we now take a string slice which starts at the item of interest, and return a string slice with the item of interest parsed. Additionally, whitespace trimming is now performed within the main read_production function, which simplifies things.